chore: add separate attributes schemas
