+++ packages/modules/attributes/lib/schemas/index.js
@@ -1,0 +1,1 @@
+export * from './attribute-set-schema';
